+++ src/core/log-plugin-version.js
