+++ src/core/options-transforms.js
