+++ src/core/options-transforms.spec.js
