+++ src/core/plugin.js
