+++ src/core/version-to-git-tag.js
