+++ src/core/version-to-git-tag.spec.js
