+++ src/dotnet.js
@@ -1,4 +1,4 @@
 const { getProjectRoot, getProjectName } = require('./dotnet-pkg-info');
-const plugin = require('./plugin');
+const plugin = require('./core/plugin');
 
 module.exports = await plugin(getProjectRoot, getProjectName);
