+++ src/npm.js
@@ -1,4 +1,4 @@
 const { getProjectRoot, getProjectName } = require('./npm-pkg-info');
-const plugin = require('./plugin');
+const plugin = require('./core/plugin');
 
 module.exports = await plugin(getProjectRoot, getProjectName);
